+++ test/cli/test_probes.py
@@ -59,8 +59,8 @@ def test_detect_null_byte_detected(openapi_30, config_factory, openapi3_base_url
     config = config_factory(base_url=openapi3_base_url, **kwargs)
     results = probes.run(openapi_30, config)
     assert results == [
-        probes.ProbeResult(
-            probe=probes.NullByteInHeader(), type=probes.ProbeResultType.FAILURE, request=ANY, response=ANY, error=None
+        probes.ProbeRun(
+            probe=probes.NullByteInHeader(), outcome=probes.ProbeOutcome.FAILURE, request=ANY, response=ANY, error=None
         )
     ]
     assert results[0].serialize() == {
@@ -78,7 +78,7 @@ def test_detect_null_byte_detected(openapi_30, config_factory, openapi3_base_url
             "uri": openapi3_base_url,
         },
         "response": None,
-        "type": "failure",
+        "outcome": "failure",
     }
 
 
@@ -116,7 +116,7 @@ def test_detect_null_byte_with_response(openapi_30, config_factory, openapi3_bas
             "status_code": 200,
             "verify": True,
         },
-        "type": "failure",
+        "outcome": "failure",
     }
 
 
@@ -124,8 +124,8 @@ def test_detect_null_byte_error(openapi_30, config_factory):
     config = config_factory(base_url="http://127.0.0.1:1")
     results = probes.run(openapi_30, config)
     assert results == [
-        probes.ProbeResult(
-            probe=probes.NullByteInHeader(), type=probes.ProbeResultType.ERROR, request=ANY, response=None, error=ANY
+        probes.ProbeRun(
+            probe=probes.NullByteInHeader(), outcome=probes.ProbeOutcome.ERROR, request=ANY, response=None, error=ANY
         )
     ]
     serialized = results[0].serialize()
@@ -152,7 +152,7 @@ def test_detect_null_byte_error(openapi_30, config_factory):
             "uri": "http://127.0.0.1:1/",
         },
         "response": None,
-        "type": "error",
+        "outcome": "error",
     }
 
 
@@ -160,8 +160,8 @@ def test_detect_null_byte_skipped(openapi_30, config_factory):
     config = config_factory(base_url=None)
     results = probes.run(openapi_30, config)
     assert results == [
-        probes.ProbeResult(
-            probe=probes.NullByteInHeader(), type=probes.ProbeResultType.SKIP, request=None, response=None, error=None
+        probes.ProbeRun(
+            probe=probes.NullByteInHeader(), outcome=probes.ProbeOutcome.SKIP, request=None, response=None, error=None
         )
     ]
     assert results[0].serialize() == {
@@ -169,5 +169,5 @@ def test_detect_null_byte_skipped(openapi_30, config_factory):
         "name": "NULL_BYTE_IN_HEADER",
         "request": None,
         "response": None,
-        "type": "skip",
+        "outcome": "skip",
     }